@@ -2,10 +2,11 @@
 
 buildscript {
     repositories {
+        google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.0.1'
+        classpath 'com.android.tools.build:gradle:4.1.0'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -14,10 +15,8 @@ buildscript {
 
 allprojects {
     repositories {
+        google()
         jcenter()
-        maven {
-            url "https://maven.google.com"
-        }
     }
 }
 
@@ -26,10 +25,10 @@ task clean(type: Delete) {
 }
 
 ext {
-    minSdkVersion = 9
-    targetSdkVersion = 26
-    compileSdkVersion = 26
-    buildToolsVersion = '27.0.3'
+    minSdkVersion = 14
+    targetSdkVersion = 30
+    compileSdkVersion = 30
+    buildToolsVersion = '30.0.2'
     sourceCompatibilityVersion = JavaVersion.VERSION_1_8
     targetCompatibilityVersion = JavaVersion.VERSION_1_8
 }
@@ -41,11 +40,11 @@ ext.deps = [
         robolectric       : 'org.robolectric:robolectric:3.0',
 
         // dependencies
-        appcompatV7       : 'com.android.support:appcompat-v7:25.4.0',
-        androidannotations: 'com.android.support:support-annotations:25.4.0',
-        design            : 'com.android.support:design:25.4.0',
+        appcompatV7       : 'androidx.appcompat:appcompat:1.0.0',
+        androidannotations: 'androidx.annotation:annotation:1.0.0',
+        design            : 'com.google.android.material:material:1.0.0',
 
-        rxandroid         : 'io.reactivex:rxandroid:1.2.1',
-        rxjava            : 'io.reactivex:rxjava:1.3.4',
+        rxandroid         : 'io.reactivex.rxjava3:rxandroid:3.0.0',
+        rxjava            : 'io.reactivex.rxjava3:rxjava:3.0.7',
         timber            : 'com.jakewharton.timber:timber:4.6.0',
 ]
