@@ -5,9 +5,11 @@ import com.hwangjr.rxbus.thread.EventThread;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 
-import rx.functions.Action1;
-import rx.subjects.PublishSubject;
-import rx.subjects.Subject;
+import io.reactivex.rxjava3.core.BackpressureStrategy;
+import io.reactivex.rxjava3.functions.Consumer;
+import io.reactivex.rxjava3.subjects.PublishSubject;
+import io.reactivex.rxjava3.subjects.Subject;
+
 
 /**
  * Wraps a single-argument 'subscriber' method on a specific object.
@@ -69,11 +71,12 @@ public class SubscriberEvent extends Event {
     }
 
     private void initObservable() {
-        subject = PublishSubject.create();
-        subject.onBackpressureBuffer().observeOn(EventThread.getScheduler(thread))
-                .subscribe(new Action1<Object>() {
+        subject = PublishSubject.create().toSerialized();
+        subject.toFlowable(BackpressureStrategy.BUFFER)
+                .observeOn(EventThread.getScheduler(thread))
+                .subscribe(new Consumer() {
                     @Override
-                    public void call(Object event) {
+                    public void accept(Object event) throws Throwable {
                         try {
                             if (valid) {
                                 handleEvent(event);
